Fix: encapsulate messages as Multipart/related.
This makes no difference in some clients like Outlook. However, Thunderbird (and
possibly other clients) insist on rigid adherence to the email RFCs, wherein the
default (Multipart/mixed) allows each item to be rendered independently. The
result is that Thunderbird just gives up when seeing a "cid:..." reference in
one of the items of a Multipart/mixed. This commit pulls us in line with the
RFCs and should ensure consistent display across clients.