Avoid exceptions terminating the current feeder.
As of this commit, an anticipated exception in a feeder is yielded
instead of raised. The effect is that the feeder can continue in the
face of an exception. The practical outcome of this is that one bad feed
can no longer interrupt the operation of its feeder on other feeds.