rss: avoid a single entry terminating a scan
This commit fixes an issue wherein a list of new entries in an RSS feed would
not get marked seen if there is any entry in the list that causes an exception.
The root cause was that an exception terminated the scan of entries without
completing its iteration and consequently updating the resources. We now still
yield exceptions, but continue our scan, meaning an exception in the middle of
list does not prevent entries later in the list being processed.