do not exit on failure to send update
Sometimes updates can fail to be sent for reasons related to the specific item
being processed (e.g. attachment that is too large). It isn't correct to
terminate the entire run at this point, so now we just skip the current update
and continue. Note that this is still reported in the logs and affects the exit
status.