suppress image embedding when exceeding 15MB
Most email hosts have some limit on the size of emails. For example, Gmail
limits you to 25MB. Instead of being oblivious to this and getting bounces, we
can stop embedding images when the current entry is growing too large. This
allows us to still receive updates that link to large images.