Decode data piped in from Git manually
It seems there is no way to handle UTF-8 decoding errors when using
universal_newlines in subprocess. Instead, the recommended way is to decode the
data separately yourself, so that is what we now do. The result of this change
is that this feeder no longer chokes on Git diffs that contain invalid UTF-8
bytes.