only attempt image embedding on first send attempt
When failing the send of an update, sometimes the failure was due to a timeout
as a result of bad network connectivity and/or a large amount of data. If we've
already failed this way once, it seems wise to try to minimise the data being
sent on our retries. We now simply ignore image embedding if we're retrying an
update.