remove redundant embed threshold check
The `downloaded` value is 0 the first time through the loop and, on any
subsequent iteration, this same check had just been done during the previous
iteration. I think this duplicate check was here under the assumption that
`downloaded` also includes the size of the content, but it doesn't. Maybe it
should.