fix: ignore UTF-8 decoding errors in 'src' URLs
It seems feeds can contain bogus bytes in this field that are only discovered
lazily when trying to access it. The simple fix seems to be just ignoring the
attempt at inlining here and leaving the image out-of-line. Perhaps email
viewers have a more clever way of dealing with these malformed bytes.